# openapi.yaml
openapi: 3.0.3
info:
  title: Banking API
  description: A simple banking API built with FastAPI
  version: 0.1.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.example.com
    description: Production server (update as needed)
...
paths:
...
  /api/v1/documents/upload:
    post:
      tags:
        - documents
      summary: Upload and process a document
      description: |
        Upload and process a document through the complete pipeline.
        
        This endpoint:
        1. Validates file size and content
        2. Determines appropriate processor based on file extension
        3. Preprocesses the document
        4. Processes and extracts structured data using Docling
        5. Handles password-protected documents (e.g., encrypted PDFs)
        6. Returns complete processing results
        
        For banking documents with sensitive information (credit cards, account numbers),
        PDFs are often password-protected. Provide the password parameter to process
        these encrypted documents.
        
        Maximum file size: 10 MB (configurable via MAX_FILE_SIZE_MB environment variable)
      operationId: upload_document
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Document file to upload and process
                password:
                  type: string
                  description: Password for encrypted/protected documents (e.g., banking PDFs with sensitive information)
                  nullable: true
      responses:
        '201':
          description: Document processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessorResponse'
        '400':
          description: Bad request - Invalid file or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '413':
          description: Request entity too large - File size exceeds maximum
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '500':
          description: Internal server error - Failed to process document
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
...
components:
  schemas:
    HTTPError:
      type: object
      properties:
        detail:
          type: string
          description: Error message
    ProcessorResponse:
      type: object
      properties:
        status:
          type: string
          description: Processing status
        processing_time_ms:
          type: number
          format: float
          description: Processing time in milliseconds
        page_count:
          type: integer
          description: Number of pages processed
        tables:
          type: array
          items:
            type: object
          description: Extracted tables from document
        data:
          type: object
          description: Extracted data including dataframes and metadata
          properties:
            dataframes:
              type: array
              items:
                type: object
            excel_file:
              type: string
              description: Path to saved Excel file with extracted tables
...